add alert to App component
